# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Blog API
  description: |

    A fast, secure blog backend built in Go.  
    - JWT authentication  
    - Image upload via Cloudinary  
    - Full CRUD (only author can edit/delete)  
    - Admin auto-created: `admin@zamrull.ng` / `admin123`
  version: 1.0.0
  contact:
    name: Eugene Okogun
    url: https://x.com/zam_rull

servers:
  - url: http://localhost:8000
    description: Local development

tags:
  - name: Auth
    description: User registration and login
  - name: Posts
    description: Blog post CRUD operations

paths:
  /register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: User created + JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Email already exists

  /login:
    post:
      tags: [Auth]
      summary: Login and get JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: JWT token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Invalid credentials

  /posts:
    get:
      tags: [Posts]
      summary: List all posts (public)
      responses:
        "200":
          description: Array of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"

    post:
      tags: [Posts]
      summary: Create a new post
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: "My Journey in Tech"
                subheading:
                  type: string
                  example: "A deep dive into my career"
                content:
                  type: string
                  example: "From Lagos to the world..."
                image:
                  type: string
                  format: binary
                  description: Optional image (JPG/PNG)
      responses:
        "200":
          description: Post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "401":
          description: Not authenticated

  /posts/{id}:
    get:
      tags: [Posts]
      summary: Get a post by ID (public)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Post ID
      responses:
        "200":
          description: Single post object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Post not found
    put:
      tags: [Posts]
      summary: Update a post (author only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Post ID
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                subheading:
                  type: string
                content:
                  type: string
                image:
                  type: string
                  format: binary
      responses:
        "200":
          description: Post updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "403":
          description: Not authorized (not author)
        "404":
          description: Post not found

    delete:
      tags: [Posts]
      summary: Delete a post (author only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Post ID
      responses:
        "200":
          description: Post deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "deleted"
        "403":
          description: Not authorized
        "404":
          description: Post not found

  /posts/{id}/video:
    post:
      tags: [Posts]
      summary: Upload one or more videos for a post
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Post ID
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                video:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Video files (MP4/MOV)
      responses:
        "200":
          description: Videos uploaded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoUploadResponse"
        "400":
          description: Invalid post ID or no video files
        "401":
          description: Not authenticated
        "403":
          description: Not authorized
        "404":
          description: Post not found
        "500":
          description: Internal server error (e.g., Cloudinary upload failure)

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    RegisterRequest:
      type: object
      required: [email, username, password]
      properties:
        email:
          type: string
          format: email
          example: "user@zamrull.ng"
        username:
          type: string
          example: "zamrull"
        password:
          type: string
          minLength: 6
          example: "pass123"

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          type: object
          properties:
            id:
              type: string
            email:
              type: string
            username:
              type: string

    TokenResponse:
      type: object
      properties:
        token:
          type: string

    Post:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        subheading:
          type: string
          nullable: true
        content:
          type: string
        image_url:
          type: string
          nullable: true
        video_url:
          type: array
          items:
            type: string
          nullable: true
        author:
          type: string
        author_id:
          type: string
        time:
          type: string
          format: date-time
        date:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    CreatePostRequest:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
          example: "My Journey in Tech"
        subheading:
          type: string
          example: "A deep dive into my career"
        content:
          type: string
          example: "From Lagos to the world..."

    VideoUploadResponse:
      type: object
      properties:
        message:
          type: string
          example: "videos uploaded successfully"
        video_urls:
          type: array
          items:
            type: string
            example: "https://res.cloudinary.com/yourcloud/video1.mp4"